Add node_modules
I normally wouldn't do this but it makes running the action much easier.
The alternative (and suggested path) is to build a distribution file and
set that as the target in action.yml. Because there are very few small
dependencies here, the package & distribution steps feel like more work
than they're worth.