Add initial version of Ruby plugin support
This is a mostly working copy from a workflow. The idea is that we see
if the plugin is being used and then we call some standard setup and
teardown functions in the main run.

The reason this is necessary and helpful is that in order to run the
prettier ruby plugin, you need to have the Ruby runtime, but you don't
need to have many gems installed.

To be on the safe side, we'll go ahead and install prettier (which is
especially necessary for >= 3.0.0). If for some reason we don't see the
gem listed or we have trouble with the Gemfile.lock, we'll log the issue
and carry on. That feels like the safe default for now, knowing there
are a number of ways to set this up. I fully expect this setup to get
more complicated as we see more use-cases. As that happens, we can get a
bit more shouty about unexpected situations. For now, let's take it
easy.

The big win here in terms of runability is moving the Gemfile out of the
way. Because the prettier ruby plugin calls `require "bundler/setup"`
during the prettier execution run, if it sees a Gemfile, bundler will
want everything in it to be installed. Moving it out of the way keeps
things moving along without having to install every gem in your project.