Setup something resembling a plugin system
I'm trying to keep this as simple as possible for now while drawing
clear lines of responsibility and keeping all of this open for future
extension.

The change itself is pretty small but what it enables is someone else to
come in and define a new plugin that exports `setup` and `teardown`
functions in addition to a `packageMatcher` regex. Once added to the
repo and added to the list of availablePlugins, it should mostly work.
Of course that assumes all the new plugin needs are these two
opportunities to do its thing, but the lines are fairly clear. And we
can expand those capabilities later if we need to.