Guard against empty list of files to fix
Because we filter changed files by a set of extensions, it's possible
that we'd get end up with an empty set of files. If this action can be
run at any time, we've got to guard against that.

In some ways, this heightens my desire to get rid of the extensions
filter. The only thing holding me back is the desire to avoid a giant
list of files that ends up causing other problems.