Remove extensions input
Prettier and its plugins already keep track of what extensions and files
they can be run on. In an effort to make sure this runs on everything it
can, we'd do well to let Prettier handle those decisions for us;
otherwise we'd be duplicating much of that functionality and knowledge.