Add autofix and append-to-file actions
The base action does the core work, but it's output isn't much use
without other actions around it. This adds the actions that are used
most with it. And being bundled in here means an app can now use
planningcenter/balto-prettier/autofix@{ref} and get all of the autofix
and commit functionality that's most useful.